Fix pip-compile output always including a local path
As we included the local package as editable, our pip compile runs always included a path. This required always manually removing the local path from the compile output afterwards and thus broke dependabot.

By reading the dependencies from a text file in setup.py, we can just reference that text file in the requirements.in and get the desired result.